split tests per domain/type
