+++ pkg/22/gmp/client/client_config_test.go
@@ -1,0 +1,61 @@
+package client
+
+import (
+	"testing"
+
+	"github.com/brennoo/go-gmp/pkg/22/gmp"
+)
+
+func TestGetConfigs(t *testing.T) {
+	cli := New(mockedConnection())
+	if cli == nil {
+		t.Fatalf("Client is nil")
+	}
+
+	cmd := &gmp.GetConfigsCommand{}
+	cmd.ConfigID = "bde773f3-2b3d-4fe6-81cb-6321ae2cc629"
+	resp, err := cli.GetConfigs(cmd)
+	if err != nil {
+		t.Fatalf("Unexpected error during GetConfigs: %s", err)
+	}
+
+	if resp.Status != "200" {
+		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
+	}
+}
+
+func TestCreateConfig(t *testing.T) {
+	cli := New(mockedConnection())
+	if cli == nil {
+		t.Fatalf("Client is nil")
+	}
+
+	cmd := &gmp.CreateConfigCommand{}
+	cmd.Name = "New Config"
+	resp, err := cli.CreateConfig(cmd)
+	if err != nil {
+		t.Fatalf("Unexpected error during CreateConfig: %s", err)
+	}
+
+	if resp.Status != "200" {
+		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
+	}
+}
+
+func TestModifyConfig(t *testing.T) {
+	cli := New(mockedConnection())
+	if cli == nil {
+		t.Fatalf("Client is nil")
+	}
+
+	cmd := &gmp.ModifyConfigCommand{}
+	cmd.Name = "Modified Config"
+	resp, err := cli.ModifyConfig(cmd)
+	if err != nil {
+		t.Fatalf("Unexpected error during ModifyConfig: %s", err)
+	}
+
+	if resp.Status != "200" {
+		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
+	}
+}
