+++ pkg/22/gmp/client/client_scanner_test.go
@@ -1,0 +1,122 @@
+package client
+
+import (
+	"testing"
+
+	"github.com/brennoo/go-gmp/pkg/22/gmp"
+)
+
+func TestCreateScanner(t *testing.T) {
+	cli := New(mockedConnection())
+	if cli == nil {
+		t.Fatalf("Client is nil")
+	}
+
+	cmd := &gmp.CreateScannerCommand{
+		Name:  "Default Scanner",
+		Host:  "localhost",
+		Port:  "9391",
+		Type:  "2",
+		CAPub: "dummy-ca-pub",
+		Credential: &gmp.CreateScannerCredential{
+			ID: "254cd3ef-bbe1-4d58-859d-21b8d0c046c6",
+		},
+	}
+	resp, err := cli.CreateScanner(cmd)
+	if err != nil {
+		t.Fatalf("Unexpected error during CreateScanner: %s", err)
+	}
+	if resp.Status != "201" {
+		t.Fatalf("Unexpected status. Expected: 201 Got: %s", resp.Status)
+	}
+	if resp.ID != "814cd30f-dee1-4d58-851d-21b8d0c048e3" {
+		t.Fatalf("Unexpected ID. Expected: 814cd30f-dee1-4d58-851d-21b8d0c048e3 Got: %s", resp.ID)
+	}
+}
+
+func TestGetScanners(t *testing.T) {
+	cli := New(mockedConnection())
+	if cli == nil {
+		t.Fatalf("Client is nil")
+	}
+
+	cmd := &gmp.GetScannersCommand{}
+	cmd.ScannerID = "ee0311e7-3247-4425-bb9c-866d59f1e0e9"
+	resp, err := cli.GetScanners(cmd)
+	if err != nil {
+		t.Fatalf("Unexpected error during GetScanners: %s", err)
+	}
+
+	if resp.Status != "200" {
+		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
+	}
+}
+
+func TestCreateScannerFail(t *testing.T) {
+	cli := New(mockedConnection())
+	if cli == nil {
+		t.Fatalf("Client is nil")
+	}
+
+	cmd := &gmp.CreateScannerCommand{
+		Name:  "",
+		Host:  "localhost",
+		Port:  "9391",
+		Type:  "2",
+		CAPub: "dummy-ca-pub",
+		Credential: &gmp.CreateScannerCredential{
+			ID: "254cd3ef-bbe1-4d58-859d-21b8d0c046c6",
+		},
+	}
+	resp, err := cli.CreateScanner(cmd)
+	if err != nil {
+		t.Fatalf("Unexpected error during CreateScanner (fail): %s", err)
+	}
+	if resp.Status != "400" {
+		t.Fatalf("Unexpected status. Expected: 400 Got: %s", resp.Status)
+	}
+}
+
+func TestModifyScanner(t *testing.T) {
+	cli := New(mockedConnection())
+	if cli == nil {
+		t.Fatalf("Client is nil")
+	}
+
+	cmd := &gmp.ModifyScannerCommand{
+		ScannerID: "scanner-uuid",
+		Name:      "Updated Scanner",
+	}
+	resp, err := cli.ModifyScanner(cmd)
+	if err != nil {
+		t.Fatalf("Unexpected error during ModifyScanner: %s", err)
+	}
+	if resp.Status != "200" {
+		t.Fatalf("Unexpected status. Expected: 200 Got: %s", resp.Status)
+	}
+	if resp.StatusText != "OK" {
+		t.Fatalf("Unexpected status text. Expected: OK Got: %s", resp.StatusText)
+	}
+}
+
+func TestModifyScannerFail(t *testing.T) {
+	cli := New(mockedConnection())
+	if cli == nil {
+		t.Fatalf("Client is nil")
+	}
+
+	cmd := &gmp.ModifyScannerCommand{
+		ScannerID: "",
+		Name:      "",
+	}
+	resp, err := cli.ModifyScanner(cmd)
+	if err != nil {
+		t.Fatalf("Unexpected error during ModifyScanner (fail): %s", err)
+	}
+	if resp.Status != "400" {
+		t.Fatalf("Unexpected status. Expected: 400 Got: %s", resp.Status)
+	}
+	if resp.StatusText != "Bad request" {
+		t.Fatalf("Unexpected status text. Expected: Bad request Got: %s", resp.StatusText)
+	}
+}
