+++ pkg/22/gmp/client/client_test.go
@@ -702,42 +702,6 @@ func TestAuthenticate(t *testing.T) {
 	}
 }
 
-func TestGetConfigs(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetConfigsCommand{}
-	cmd.ConfigID = "bde773f3-2b3d-4fe6-81cb-6321ae2cc629"
-	resp, err := cli.GetConfigs(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetConfigs: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestGetScanners(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetScannersCommand{}
-	cmd.ScannerID = "ee0311e7-3247-4425-bb9c-866d59f1e0e9"
-	resp, err := cli.GetScanners(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetScanners: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
 func TestGetPreferences(t *testing.T) {
 	cli := New(mockedConnection())
 	if cli == nil {
@@ -749,260 +713,6 @@ func TestGetPreferences(t *testing.T) {
 	resp, err := cli.GetPreferences(cmd)
 	if err != nil {
 		t.Fatalf("Unexpected error during GetPreferences: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestCreateConfig(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.CreateConfigCommand{}
-	cmd.Name = "New Config"
-	resp, err := cli.CreateConfig(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateConfig: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestModifyConfig(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.ModifyConfigCommand{}
-	cmd.Name = "Modified Config"
-	resp, err := cli.ModifyConfig(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyConfig: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestCreateTask(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.CreateTaskCommand{}
-	cmd.Name = "New Task"
-	resp, err := cli.CreateTask(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateTask: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestModifyTask(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.ModifyTaskCommand{}
-	cmd.TaskID = "e512e2ca-9d0e-4bf3-bc73-7fbe6e9bbf31"
-	cmd.Comment = "Modified Task Comment"
-	resp, err := cli.ModifyTask(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyTask: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestCreateTarget(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.CreateTargetCommand{}
-	cmd.Name = "New Target"
-	resp, err := cli.CreateTarget(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateTarget: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestStartTask(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.StartTaskCommand{}
-	cmd.TaskID = "e512e2ca-9d0e-4bf3-bc73-7fbe6e9bbf31"
-	resp, err := cli.StartTask(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during StartTask: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestGetTasks(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetTasksCommand{}
-	cmd.TaskID = "e512e2ca-9d0e-4bf3-bc73-7fbe6e9bbf31"
-	resp, err := cli.GetTasks(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetTasks: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestStopTask(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.StopTaskCommand{}
-	cmd.TaskID = "e512e2ca-9d0e-4bf3-bc73-7fbe6e9bbf31"
-	resp, err := cli.StopTask(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during StopTask: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestDeleteTask(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.DeleteTaskCommand{}
-	cmd.TaskID = "e512e2ca-9d0e-4bf3-bc73-7fbe6e9bbf31"
-	resp, err := cli.DeleteTask(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteTask: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestGetResults(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetResultsCommand{}
-	cmd.TaskID = "e512e2ca-9d0e-4bf3-bc73-7fbe6e9bbf31"
-	resp, err := cli.GetResults(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetResults: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestGetVulns(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetVulnsCommand{}
-	cmd.VulnID = "1.3.6.1.4.1.25623.1.0.808160"
-	resp, err := cli.GetVulns(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetVulns: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestModifyTarget(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.ModifyTargetCommand{}
-	cmd.TargetID = "254cd3ef-bbe1-4d58-859d-21b8d0c046c6"
-	cmd.Name = "Modified Target"
-	resp, err := cli.ModifyTarget(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyTarget: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestGetTargets(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetTargetsCommand{}
-	cmd.TargetID = "254cd3ef-bbe1-4d58-859d-21b8d0c046c6"
-	resp, err := cli.GetTargets(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetTargets: %s", err)
-	}
-
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestDeleteTarget(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.DeleteTargetCommand{}
-	cmd.TargetID = "254cd3ef-bbe1-4d58-859d-21b8d0c046c6"
-	resp, err := cli.DeleteTarget(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteTarget: %s", err)
 	}
 
 	if resp.Status != "200" {
@@ -1025,799 +735,6 @@ func TestGetPortLists(t *testing.T) {
 
 	if resp.Status != "200" {
 		t.Fatalf("Unexpected status. \nExpected: 200 \nGot: %s", resp.Status)
-	}
-}
-
-func TestCreateAlert(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.CreateAlertRequest{}
-	cmd.Name = "Test Alert"
-	cmd.Condition = gmp.AlertCondition{
-		Text: "Severity at least",
-		Data: []gmp.AlertData{
-			{Name: "severity", Text: "5.5"},
-		},
-	}
-	cmd.Event = gmp.AlertEvent{
-		Text: "Task run status changed",
-		Data: []gmp.AlertData{
-			{Name: "status", Text: "Done"},
-		},
-	}
-	cmd.Method = gmp.AlertMethod{
-		Text: "Email",
-		Data: []gmp.AlertData{
-			{Name: "to_address", Text: "test@example.org"},
-			{Name: "from_address", Text: "alert@example.org"},
-		},
-	}
-	resp, err := cli.CreateAlert(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateAlert: %s", err)
-	}
-
-	if resp.Status != "201" {
-		t.Fatalf("Unexpected status. \nExpected: 201 \nGot: %s", resp.Status)
-	}
-
-	if resp.StatusText != "OK, resource created" {
-		t.Fatalf("Unexpected status text. \nExpected: OK, resource created \nGot: %s", resp.StatusText)
-	}
-
-	if resp.ID != "254cd3ef-bbe1-4d58-859d-21b8d0c046c6" {
-		t.Fatalf("Unexpected ID. \nExpected: 254cd3ef-bbe1-4d58-859d-21b8d0c046c6 \nGot: %s", resp.ID)
-	}
-}
-
-func TestGetAlerts(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetAlertsCommand{}
-	resp, err := cli.GetAlerts(cmd)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-
-	if resp.Status != "200" {
-		t.Errorf("expected status '200', got %s", resp.Status)
-	}
-}
-
-func TestModifyAlert(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	alertID := "914b59f8-25f5-4c8f-832c-2379cd625236"
-	name := "Updated Alert Name"
-	comment := "Updated comment"
-	filter := &gmp.ModifyAlertFilter{ID: "7a06bd00-7e4a-4669-b7d7-8fe65ec64a41"}
-	cmd := &gmp.ModifyAlertCommand{
-		AlertID: alertID,
-		Name:    &name,
-		Comment: &comment,
-		Filter:  filter,
-	}
-	resp, err := cli.ModifyAlert(cmd)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("expected status '200', got %s", resp.Status)
-	}
-}
-
-func TestDeleteAlert(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	alertID := "267a3405-e84a-47da-97b2-5fa0d2e8995e"
-	cmd := &gmp.DeleteAlertCommand{
-		AlertID:  alertID,
-		Ultimate: true,
-	}
-	resp, err := cli.DeleteAlert(cmd)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("expected status '200', got %s", resp.Status)
-	}
-}
-
-func TestTestAlert(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	alertID := "97390ade-e075-11df-9973-002264764cea"
-	cmd := &gmp.TestAlertCommand{
-		AlertID: alertID,
-	}
-	resp, err := cli.TestAlert(cmd)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("expected status '200', got %s", resp.Status)
-	}
-}
-
-func TestResumeTask(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	taskID := "267a3405-e84a-47da-97b2-5fa0d2e8995e"
-	cmd := &gmp.ResumeTaskCommand{
-		TaskID: taskID,
-	}
-	resp, err := cli.ResumeTask(cmd)
-	if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("expected status '200', got %s", resp.Status)
-	}
-	if resp.ReportID != "330ee785-c2c0-4d4c-ab96-725142c9b789" {
-		t.Errorf("expected report_id '330ee785-c2c0-4d4c-ab96-725142c9b789', got %s", resp.ReportID)
-	}
-}
-
-func TestCreateAsset(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.CreateAssetCommand{
-		Asset: &gmp.CreateAssetAsset{
-			Name:    "Localhost",
-			Type:    "host",
-			Comment: "Test asset",
-		},
-	}
-	resp, err := cli.CreateAsset(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateAsset: %s", err)
-	}
-	if resp.Status != "201" {
-		t.Fatalf("Unexpected status. Expected: 201 Got: %s", resp.Status)
-	}
-	if resp.StatusText != "OK, resource created" {
-		t.Fatalf("Unexpected status text. Expected: OK, resource created Got: %s", resp.StatusText)
-	}
-	if resp.ID != "254cd3ef-bbe1-4d58-859d-21b8d0c046c6" {
-		t.Fatalf("Unexpected ID. Expected: 254cd3ef-bbe1-4d58-859d-21b8d0c046c6 Got: %s", resp.ID)
-	}
-}
-
-func TestCreateAssetWithReport(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.CreateAssetCommand{
-		Report: &gmp.CreateAssetReport{
-			ID: "report-uuid",
-			Filter: &gmp.CreateAssetReportFilter{
-				Term: "min_qod=70",
-			},
-		},
-	}
-	resp, err := cli.CreateAsset(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateAsset (report): %s", err)
-	}
-	if resp.Status != "201" {
-		t.Fatalf("Unexpected status. Expected: 201 Got: %s", resp.Status)
-	}
-	if resp.StatusText != "OK, resource created" {
-		t.Fatalf("Unexpected status text. Expected: OK, resource created Got: %s", resp.StatusText)
-	}
-	if resp.ID != "report-asset-uuid" {
-		t.Fatalf("Unexpected ID. Expected: report-asset-uuid Got: %s", resp.ID)
-	}
-}
-
-func TestModifyAsset(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.ModifyAssetCommand{
-		AssetID: "914b59f8-25f5-4c8f-832c-2379cd625236",
-		Comment: "New comment",
-	}
-	resp, err := cli.ModifyAsset(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyAsset: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. Expected: 200 Got: %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Fatalf("Unexpected status text. Expected: OK Got: %s", resp.StatusText)
-	}
-}
-
-func TestGetAssets(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetAssetsCommand{
-		AssetID: "b493b7a8-7489-11df-a3ec-002264764cea",
-	}
-	resp, err := cli.GetAssets(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetAssets: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. Expected: 200 Got: %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Fatalf("Unexpected status text. Expected: OK Got: %s", resp.StatusText)
-	}
-	if len(resp.Assets) != 1 {
-		t.Fatalf("Expected 1 asset, got %d", len(resp.Assets))
-	}
-	asset := resp.Assets[0]
-	if asset.ID != "b493b7a8-7489-11df-a3ec-002264764cea" {
-		t.Fatalf("Unexpected asset ID. Expected: b493b7a8-7489-11df-a3ec-002264764cea Got: %s", asset.ID)
-	}
-	if asset.Name != "Localhost" {
-		t.Fatalf("Unexpected asset name. Expected: Localhost Got: %s", asset.Name)
-	}
-}
-
-func TestDeleteAsset(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.DeleteAssetCommand{
-		AssetID: "267a3405-e84a-47da-97b2-5fa0d2e8995e",
-	}
-	resp, err := cli.DeleteAsset(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteAsset: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Fatalf("Unexpected status. Expected: 200 Got: %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Fatalf("Unexpected status text. Expected: OK Got: %s", resp.StatusText)
-	}
-}
-
-func TestCreateSchedule(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.CreateScheduleCommand{
-		Name:      "Monthly Scan",
-		Timezone:  "America/New_York",
-		ICalendar: "BEGIN:VCALENDAR\nVERSION:2.0\nPRODID:-//XXX//NONSGML//EN\nBEGIN:VEVENT\nDTSTART;TZID=\"America/New_York\":20221214T000100\nRRULE:FREQ=MONTHLY;BYDAY=3WE;\nEND:VEVENT\nEND:VCALENDAR",
-	}
-	resp, err := cli.CreateSchedule(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateSchedule: %s", err)
-	}
-	if resp.Status != "201" {
-		t.Errorf("Expected status 201, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK, resource created" {
-		t.Errorf("Expected status text 'OK, resource created', got '%s'", resp.StatusText)
-	}
-	if resp.ID != "254cd3ef-bbe1-4d58-859d-21b8d0c046c6" {
-		t.Errorf("Expected ID '254cd3ef-bbe1-4d58-859d-21b8d0c046c6', got '%s'", resp.ID)
-	}
-}
-
-func TestModifySchedule(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.ModifyScheduleCommand{
-		ScheduleID: "254cd3ef-bbe1-4d58-859d-21b8d0c046c6",
-		Name:       "Weekly Scan",
-	}
-	resp, err := cli.ModifySchedule(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifySchedule: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-	}
-}
-
-func TestGetSchedules(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetSchedulesCommand{
-		ScheduleID: "ddda859a-45be-4c58-85b3-517c66230232",
-	}
-	resp, err := cli.GetSchedules(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetSchedules: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-	}
-	if len(resp.Schedules) != 1 {
-		t.Errorf("Expected 1 schedule, got %d", len(resp.Schedules))
-	}
-	if resp.Schedules[0].Name != "Every day" {
-		t.Errorf("Expected schedule name 'Every day', got '%s'", resp.Schedules[0].Name)
-	}
-	if resp.Schedules[0].Timezone != "UTC" {
-		t.Errorf("Expected timezone 'UTC', got '%s'", resp.Schedules[0].Timezone)
-	}
-}
-
-func TestDeleteSchedule(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.DeleteScheduleCommand{
-		ScheduleID: "267a3405-e84a-47da-97b2-5fa0d2e8995e",
-		Ultimate:   "0",
-	}
-	resp, err := cli.DeleteSchedule(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteSchedule: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-	}
-}
-
-func TestCreateOverride(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.CreateOverrideCommand{
-		Text:      "This is actually of little concern.",
-		NVT:       gmp.CreateOverrideNVT{OID: "1.3.6.1.4.1.25623.1.0.10330"},
-		NewThreat: "Low",
-		Result:    "254cd3ef-bbe1-4d58-859d-21b8d0c046c6",
-	}
-	resp, err := cli.CreateOverride(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateOverride: %s", err)
-	}
-	if resp.Status != "201" {
-		t.Errorf("Expected status 201, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK, resource created" {
-		t.Errorf("Expected status text 'OK, resource created', got '%s'", resp.StatusText)
-	}
-	if resp.ID != "254cd3ef-bbe1-4d58-859d-21b8d0c046c6" {
-		t.Errorf("Expected ID '254cd3ef-bbe1-4d58-859d-21b8d0c046c6', got '%s'", resp.ID)
-	}
-}
-
-func TestDeleteOverride(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.DeleteOverrideCommand{
-		OverrideID: "267a3405-e84a-47da-97b2-5fa0d2e8995e",
-		Ultimate:   "0",
-	}
-	resp, err := cli.DeleteOverride(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteOverride: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text OK, got %s", resp.StatusText)
-	}
-}
-
-func TestGetOverrides(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.GetOverridesCommand{
-		OverrideID: "b76b81a7-9df8-42df-afff-baa9d4620128",
-	}
-	resp, err := cli.GetOverrides(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetOverrides: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text OK, got %s", resp.StatusText)
-	}
-	if len(resp.Overrides) != 1 {
-		t.Errorf("Expected 1 override, got %d", len(resp.Overrides))
-	}
-	if resp.Overrides[0].ID != "b76b81a7-9df8-42df-afff-baa9d4620128" {
-		t.Errorf("Expected override ID b76b81a7-9df8-42df-afff-baa9d4620128, got %s", resp.Overrides[0].ID)
-	}
-}
-
-func TestModifyOverride(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	cmd := &gmp.ModifyOverrideCommand{
-		OverrideID: "254cd3ef-bbe1-4d58-859d-21b8d0c046c6",
-		Text:       "This issue is less important in our setup.",
-		NewThreat:  "Low",
-	}
-	resp, err := cli.ModifyOverride(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyOverride: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text OK, got %s", resp.StatusText)
-	}
-}
-
-func TestCreateReport(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.CreateReportCommand{
-		Report: &gmp.ReportWrapper{
-			ID:          "report-uuid",
-			FormatID:    "format-uuid",
-			Extension:   "xml",
-			ContentType: "text/xml",
-		},
-		Task: &gmp.CreateReportTask{
-			ID: "task-uuid",
-		},
-		InAssets: nil,
-	}
-	resp, err := cli.CreateReport(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateReport: %s", err)
-	}
-	if resp.Status != "201" {
-		t.Errorf("Expected status 201, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK, resource created" {
-		t.Errorf("Expected status text 'OK, resource created', got '%s'", resp.StatusText)
-	}
-	if resp.ID != "created-report-id" {
-		t.Errorf("Expected ID 'created-report-id', got '%s'", resp.ID)
-	}
-
-	// Failure case
-	cmdFail := &gmp.CreateReportCommand{
-		Report: &gmp.ReportWrapper{
-			ID: "wrong-id",
-		},
-		Task: &gmp.CreateReportTask{
-			ID: "wrong-task",
-		},
-	}
-	respFail, err := cli.CreateReport(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateReport (fail): %s", err)
-	}
-	if respFail.Status != "400" {
-		t.Errorf("Expected status 400, got %s", respFail.Status)
-	}
-	if respFail.StatusText != "Bad request" {
-		t.Errorf("Expected status text 'Bad request', got '%s'", respFail.StatusText)
-	}
-}
-
-func TestGetReports(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.GetReportsCommand{
-		ReportID:    "report-uuid",
-		FormatID:    "format-uuid",
-		Extension:   "xml",
-		ContentType: "text/xml",
-	}
-	resp, err := cli.GetReports(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetReports: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-	}
-	if len(resp.Reports) != 1 {
-		t.Errorf("Expected 1 report, got %d", len(resp.Reports))
-	} else {
-		report := resp.Reports[0]
-		if report.ID != "report-uuid" {
-			t.Errorf("Expected report ID 'report-uuid', got '%s'", report.ID)
-		}
-		if report.FormatID != "format-uuid" {
-			t.Errorf("Expected format ID 'format-uuid', got '%s'", report.FormatID)
-		}
-		if report.Extension != "xml" {
-			t.Errorf("Expected extension 'xml', got '%s'", report.Extension)
-		}
-		if report.ContentType != "text/xml" {
-			t.Errorf("Expected content type 'text/xml', got '%s'", report.ContentType)
-		}
-	}
-
-	// Failure case
-	cmdFail := &gmp.GetReportsCommand{
-		ReportID: "wrong-id",
-	}
-	respFail, err := cli.GetReports(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetReports (fail): %s", err)
-	}
-	if respFail.Status != "404" {
-		t.Errorf("Expected status 404, got %s", respFail.Status)
-	}
-	if respFail.StatusText != "Not found" {
-		t.Errorf("Expected status text 'Not found', got '%s'", respFail.StatusText)
-	}
-}
-
-func TestDeleteReport(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.DeleteReportCommand{
-		ReportID: "report-uuid",
-	}
-	resp, err := cli.DeleteReport(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteReport: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-	}
-
-	// Failure case
-	cmdFail := &gmp.DeleteReportCommand{
-		ReportID: "wrong-id",
-	}
-	respFail, err := cli.DeleteReport(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteReport (fail): %s", err)
-	}
-	if respFail.Status != "404" {
-		t.Errorf("Expected status 404, got %s", respFail.Status)
-	}
-	if respFail.StatusText != "Not found" {
-		t.Errorf("Expected status text 'Not found', got '%s'", respFail.StatusText)
-	}
-}
-
-func TestCreateReportFormat(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.CreateReportFormatCommand{
-		Name: "Test Format",
-		Copy: "copy-uuid",
-	}
-	resp, err := cli.CreateReportFormat(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateReportFormat: %s", err)
-	}
-	if resp.Status != "201" {
-		t.Errorf("Expected status 201, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK, resource created" {
-		t.Errorf("Expected status text 'OK, resource created', got '%s'", resp.StatusText)
-	}
-	if resp.ID != "created-format-id" {
-		t.Errorf("Expected ID 'created-format-id', got '%s'", resp.ID)
-	}
-
-	// Failure case
-	cmdFail := &gmp.CreateReportFormatCommand{
-		Name: "Wrong Format",
-		Copy: "wrong-copy",
-	}
-	respFail, err := cli.CreateReportFormat(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateReportFormat (fail): %s", err)
-	}
-	if respFail.Status != "400" {
-		t.Errorf("Expected status 400, got %s", respFail.Status)
-	}
-	if respFail.StatusText != "Bad request" {
-		t.Errorf("Expected status text 'Bad request', got '%s'", respFail.StatusText)
-	}
-}
-
-func TestModifyReportFormat(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case: modify name
-	cmd := &gmp.ModifyReportFormatCommand{
-		ReportFormatID: "format-uuid",
-		Name:           "New Name",
-	}
-	resp, err := cli.ModifyReportFormat(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyReportFormat: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-	}
-
-	// Success case: modify param
-	cmdParam := &gmp.ModifyReportFormatCommand{
-		ReportFormatID: "format-uuid",
-		Param: &gmp.ModifyReportFormatParam{
-			Name:  "Background Colour",
-			Value: "red",
-		},
-	}
-	respParam, err := cli.ModifyReportFormat(cmdParam)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyReportFormat (param): %s", err)
-	}
-	if respParam.Status != "200" {
-		t.Errorf("Expected status 200, got %s", respParam.Status)
-	}
-	if respParam.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", respParam.StatusText)
-	}
-
-	// Failure case
-	cmdFail := &gmp.ModifyReportFormatCommand{
-		ReportFormatID: "wrong-uuid",
-		Name:           "Invalid",
-	}
-	respFail, err := cli.ModifyReportFormat(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyReportFormat (fail): %s", err)
-	}
-	if respFail.Status != "400" {
-		t.Errorf("Expected status 400, got %s", respFail.Status)
-	}
-	if respFail.StatusText != "Bad request" {
-		t.Errorf("Expected status text 'Bad request', got '%s'", respFail.StatusText)
-	}
-}
-
-func TestGetReportConfigs(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.GetReportConfigsCommand{
-		ReportConfigID: "config-uuid",
-	}
-	resp, err := cli.GetReportConfigs(cmd)
-	if err != nil {
-		t.Fatalf("GetReportConfigs failed: %v", err)
-	}
-	if resp.Status != "200" || resp.StatusText != "OK" || len(resp.ReportConfigs) != 1 {
-		t.Errorf("unexpected response: %+v", resp)
-	}
-	if resp.ReportConfigs[0].ID != "config-uuid" || resp.ReportConfigs[0].Name != "Test config" {
-		t.Errorf("unexpected report config: %+v", resp.ReportConfigs[0])
-	}
-
-	// Failure case
-	cmd = &gmp.GetReportConfigsCommand{
-		ReportConfigID: "notfound",
-	}
-	resp, err = cli.GetReportConfigs(cmd)
-	if err != nil {
-		t.Fatalf("GetReportConfigs failed: %v", err)
-	}
-	if resp.Status != "404" || resp.StatusText != "Not found" {
-		t.Errorf("unexpected response: %+v", resp)
-	}
-}
-
-func TestDeleteReportConfig(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.DeleteReportConfigCommand{
-		ReportConfigID: "config-uuid",
-		Ultimate:       "1",
-	}
-	resp, err := cli.DeleteReportConfig(cmd)
-	if err != nil {
-		t.Fatalf("DeleteReportConfig failed: %v", err)
-	}
-	if resp.Status != "200" || resp.StatusText != "OK" {
-		t.Errorf("unexpected response: %+v", resp)
-	}
-
-	// Failure case
-	cmd = &gmp.DeleteReportConfigCommand{
-		ReportConfigID: "notfound",
-		Ultimate:       "1",
-	}
-	resp, err = cli.DeleteReportConfig(cmd)
-	if err != nil {
-		t.Fatalf("DeleteReportConfig failed: %v", err)
-	}
-	if resp.Status != "404" || resp.StatusText != "Not found" {
-		t.Errorf("unexpected response: %+v", resp)
 	}
 }
 
@@ -1870,256 +787,4 @@ func TestGetSystemReports(t *testing.T) {
 	if resp.Status != "404" || resp.StatusText != "Not found" {
 		t.Errorf("unexpected response: %+v", resp)
 	}
-}
-
-func TestCreateCredential(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.CreateCredentialCommand{
-		Name: "Test Credential",
-	}
-	resp, err := cli.CreateCredential(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateCredential: %s", err)
-	}
-	if resp.Status != "201" {
-		t.Errorf("Expected status 201, got %s", resp.Status)
-	}
-	if resp.ID != "created-credential-id" {
-		t.Errorf("Expected ID 'created-credential-id', got %s", resp.ID)
-	}
-
-	// Failure case
-	cmdFail := &gmp.CreateCredentialCommand{
-		Name: "",
-	}
-	respFail, err := cli.CreateCredential(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during CreateCredential (fail): %s", err)
-	}
-	if respFail.Status != "400" {
-		t.Errorf("Expected status 400, got %s", respFail.Status)
-	}
-}
-
-func TestModifyCredential(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.ModifyCredentialCommand{
-		CredentialID: "cred-uuid",
-		Name:         "Updated Credential",
-	}
-	resp, err := cli.ModifyCredential(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyCredential: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-	}
-
-	// Failure case
-	cmdFail := &gmp.ModifyCredentialCommand{
-		CredentialID: "",
-	}
-	respFail, err := cli.ModifyCredential(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during ModifyCredential (fail): %s", err)
-	}
-	if respFail.Status != "400" {
-		t.Errorf("Expected status 400, got %s", respFail.Status)
-	}
-	if respFail.StatusText != "Bad request" {
-		t.Errorf("Expected status text 'Bad request', got '%s'", respFail.StatusText)
-	}
-}
-
-func TestGetCredentials(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// List all credentials
-	cmd := &gmp.GetCredentialsCommand{}
-	resp, err := cli.GetCredentials(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetCredentials: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if len(resp.Credentials) != 2 {
-		t.Errorf("Expected 2 credentials, got %d", len(resp.Credentials))
-	}
-	if resp.Credentials[0].ID != "cred-uuid-1" || resp.Credentials[1].ID != "cred-uuid-2" {
-		t.Errorf("Unexpected credential IDs: %+v", resp.Credentials)
-	}
-
-	// Fetch single credential
-	cmdSingle := &gmp.GetCredentialsCommand{CredentialID: "cred-uuid-1"}
-	respSingle, err := cli.GetCredentials(cmdSingle)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetCredentials (single): %s", err)
-	}
-	if respSingle.Status != "200" {
-		t.Errorf("Expected status 200, got %s", respSingle.Status)
-	}
-	if len(respSingle.Credentials) != 1 || respSingle.Credentials[0].ID != "cred-uuid-1" {
-		t.Errorf("Unexpected credential: %+v", respSingle.Credentials)
-	}
-
-	// Failure case
-	cmdFail := &gmp.GetCredentialsCommand{CredentialID: "notfound"}
-	respFail, err := cli.GetCredentials(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during GetCredentials (fail): %s", err)
-	}
-	if respFail.Status != "404" {
-		t.Errorf("Expected status 404, got %s", respFail.Status)
-	}
-}
-
-func TestDeleteCredential(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	// Success case
-	cmd := &gmp.DeleteCredentialCommand{
-		CredentialID: "cred-uuid-1",
-		Ultimate:     "1",
-	}
-	resp, err := cli.DeleteCredential(cmd)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteCredential: %s", err)
-	}
-	if resp.Status != "200" {
-		t.Errorf("Expected status 200, got %s", resp.Status)
-	}
-	if resp.StatusText != "OK" {
-		t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-	}
-
-	// Failure case
-	cmdFail := &gmp.DeleteCredentialCommand{
-		CredentialID: "notfound",
-		Ultimate:     "1",
-	}
-	respFail, err := cli.DeleteCredential(cmdFail)
-	if err != nil {
-		t.Fatalf("Unexpected error during DeleteCredential (fail): %s", err)
-	}
-	if respFail.Status != "404" {
-		t.Errorf("Expected status 404, got %s", respFail.Status)
-	}
-	if respFail.StatusText != "Not found" {
-		t.Errorf("Expected status text 'Not found', got '%s'", respFail.StatusText)
-	}
-}
-
-func runScannerTest(t *testing.T, name string, run func(t *testing.T)) {
-	t.Run(name, run)
-}
-
-func TestCreateScanner(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	runScannerTest(t, "success", func(t *testing.T) {
-		cmd := &gmp.CreateScannerCommand{
-			Name:  "Default Scanner",
-			Host:  "localhost",
-			Port:  "9391",
-			Type:  "2",
-			CAPub: "dummy-ca-pub",
-			Credential: &gmp.CreateScannerCredential{
-				ID: "254cd3ef-bbe1-4d58-859d-21b8d0c046c6",
-			},
-		}
-		resp, err := cli.CreateScanner(cmd)
-		if err != nil {
-			t.Fatalf("Unexpected error during CreateScanner: %s", err)
-		}
-		if resp.Status != "201" {
-			t.Errorf("Expected status 201, got %s", resp.Status)
-		}
-		if resp.ID != "814cd30f-dee1-4d58-851d-21b8d0c048e3" {
-			t.Errorf("Expected ID '814cd30f-dee1-4d58-851d-21b8d0c048e3', got %s", resp.ID)
-		}
-	})
-
-	runScannerTest(t, "fail", func(t *testing.T) {
-		cmdFail := &gmp.CreateScannerCommand{
-			Name:  "",
-			Host:  "localhost",
-			Port:  "9391",
-			Type:  "2",
-			CAPub: "dummy-ca-pub",
-			Credential: &gmp.CreateScannerCredential{
-				ID: "254cd3ef-bbe1-4d58-859d-21b8d0c046c6",
-			},
-		}
-		respFail, err := cli.CreateScanner(cmdFail)
-		if err != nil {
-			t.Fatalf("Unexpected error during CreateScanner (fail): %s", err)
-		}
-		if respFail.Status != "400" {
-			t.Errorf("Expected status 400, got %s", respFail.Status)
-		}
-	})
-}
-
-func TestModifyScanner(t *testing.T) {
-	cli := New(mockedConnection())
-	if cli == nil {
-		t.Fatalf("Client is nil")
-	}
-
-	runScannerTest(t, "success", func(t *testing.T) {
-		cmd := &gmp.ModifyScannerCommand{
-			ScannerID: "scanner-uuid",
-			Name:      "Updated Scanner",
-		}
-		resp, err := cli.ModifyScanner(cmd)
-		if err != nil {
-			t.Fatalf("Unexpected error during ModifyScanner: %s", err)
-		}
-		if resp.Status != "200" {
-			t.Errorf("Expected status 200, got %s", resp.Status)
-		}
-		if resp.StatusText != "OK" {
-			t.Errorf("Expected status text 'OK', got '%s'", resp.StatusText)
-		}
-	})
-
-	runScannerTest(t, "fail", func(t *testing.T) {
-		cmdFail := &gmp.ModifyScannerCommand{
-			ScannerID: "",
-			Name:      "",
-		}
-		respFail, err := cli.ModifyScanner(cmdFail)
-		if err != nil {
-			t.Fatalf("Unexpected error during ModifyScanner (fail): %s", err)
-		}
-		if respFail.Status != "400" {
-			t.Errorf("Expected status 400, got %s", respFail.Status)
-		}
-		if respFail.StatusText != "Bad request" {
-			t.Errorf("Expected status text 'Bad request', got '%s'", respFail.StatusText)
-		}
-	})
 }
